Refactored for better incapsulation and seperaion of concerns
